move precise_nano_sleep into util
