+++ selfdrive/ui/replay/util.cc
@@ -4,6 +4,7 @@
 #include <bzlib.h>
 #include <curl/curl.h>
 
+#include "selfdrive/common/timing.h"
 #include "selfdrive/common/util.h"
 
 struct CURLGlobalInitializer {
@@ -129,4 +130,22 @@ bool readBZ2File(const std::string_view file, std::ostream &stream) {
   bool success = (bzerror == BZ_STREAM_END);
   BZ2_bzReadClose(&bzerror, bz_file);
   return success;
+}
+
+void precise_nano_sleep(long sleep_ns) {
+  const long estimate_ns = 1 * 1e6;  // 1ms
+  struct timespec req = {.tv_nsec = estimate_ns};
+  uint64_t start_sleep = nanos_since_boot();
+  while (sleep_ns > estimate_ns) {
+    nanosleep(&req, nullptr);
+    uint64_t end_sleep = nanos_since_boot();
+    sleep_ns -= (end_sleep - start_sleep);
+    start_sleep = end_sleep;
+  }
+  // spin wait
+  if (sleep_ns > 0) {
+    while ((nanos_since_boot() - start_sleep) <= sleep_ns) { 
+      usleep(0);
+    }
+  }
 }
